+++ WooCommerce/WooCommerceTests/Mockups/MockImageDownloader.swift
@@ -1,34 +1,29 @@
-@testable import Kingfisher
+import UIKit
+@testable import WooCommerce
 
-final class MockKingfisherImageDownloader: Kingfisher.ImageDownloader {
+final class MockImageDownloadTask: ImageDownloadTask {
+    private(set) var isCancelled: Bool = false
+
+    func cancel() {
+        isCancelled = true
+    }
+}
+
+final class MockImageDownloader: ImageDownloader {
     // Mocks in-memory cache.
     private let imagesByKey: [String: UIImage]
 
     init(imagesByKey: [String: UIImage]) {
         self.imagesByKey = imagesByKey
-        super.init(name: "Mock!")
     }
 
-    override func downloadImage(with url: URL,
-                                options: KingfisherOptionsInfo? = nil,
-                                progressBlock: DownloadProgressBlock?,
-                                completionHandler: ((Result<ImageLoadingResult, KingfisherError>) -> Void)? = nil) -> DownloadTask? {
+    func downloadImage(with url: URL, onCompletion: ((Result<UIImage, Error>) -> Void)?) -> ImageDownloadTask? {
         if let image = imagesByKey[url.absoluteString] {
-            completionHandler?(.success(.init(image: image, url: url, originalData: Data())))
+            onCompletion?(.success(image))
         } else {
-            completionHandler?(.failure(.cacheError(reason: .imageNotExisting(key: url.absoluteString) )))
+            let error = NSError(domain: "MockDownloadable", code: 1, userInfo: nil)
+            onCompletion?(.failure(error))
         }
-        return nil
-    }
-
-    override func downloadImage(with url: URL,
-                                options: KingfisherParsedOptionsInfo,
-                                completionHandler: ((Result<ImageLoadingResult, KingfisherError>) -> Void)? = nil) -> DownloadTask? {
-        if let image = imagesByKey[url.absoluteString] {
-            completionHandler?(.success(.init(image: image, url: url, originalData: Data())))
-        } else {
-            completionHandler?(.failure(.cacheError(reason: .imageNotExisting(key: url.absoluteString) )))
-        }
-        return nil
+        return MockImageDownloadTask()
     }
 }
